requirements.txt
flask==1.1.1
PyMySQL==0.9.3
redis==3.3.11
gevent==1.4.0
pymongo==3.10.0
pika==1.1.0
schedule==0.6.0
pyyaml==5.2
requests==2.22.0
numpy==1.18.0
six==1.13.0
py2neo==4.3.0